- Fix token time data type fatal error

+++ ChatApplication/Settings/SettingViewModel.swift
@@ -15,13 +15,13 @@ class SettingViewModel: ObservableObject {
     var model = SettingModel()
     
     @Published
-    var secondToExpire:Int = 0
+    var secondToExpire:Double = 0
     
     func startTokenTimer(){
         Timer.scheduledTimer(withTimeInterval: 1, repeats: true) { [weak self] timer in
             if let createDate = TokenManager.shared.getCreateTokenDate() , let ssoTokenExipreTime = TokenManager.shared.getSSOTokenFromUserDefaults()?.expiresIn{
                 let expireIn = createDate.advanced(by:  Double(ssoTokenExipreTime)).timeIntervalSince1970 - Date().timeIntervalSince1970
-                self?.secondToExpire = Int(expireIn)
+                self?.secondToExpire = Double(expireIn)
             }
         }
     }
